Fix reject road speed calculation precision

+++ source/game/kart/KartReject.cc
@@ -157,7 +157,9 @@ bool KartReject::calcRejection() {
             colInfo.tangentOff += worldPos;
 
             f32 yOffset = bsp().initialYPos * scale().y;
-            f32 speedScalar = bVar15 ? 1.0f : EGG::Mathf::abs(speed()) * 0.01f - 0.3f;
+            f32 speedScalar = bVar15 ?
+                    1.0f :
+                    static_cast<f32>(static_cast<f64>(EGG::Mathf::abs(speed()) * 0.01f) - 0.3d);
             speedScalar = std::min(1.0f, std::max(0.0f, speedScalar));
 
             EGG::Vector3f posOffset =
